Added a few quaternion unit tests, and removed a few mixed-precision methods (I want to remove them, because you shouldn't really mix them without really thinking about it); I'm also contemplating removing all transform-Non-Unit-functions, because quaternions shall be normalized...

+++ OpenXR/src/me/anno/openxr/OpenXRActions.kt
@@ -343,7 +343,7 @@ class OpenXRActions(val instance: XrInstance, val session: XrSession, identityPo
                 .add(additionalOffset)
             controller.rotation
                 .set(additionalRotation)
-                .mul(rot.x(), rot.y(), rot.z(), rot.w())
+                .mul(rot.x().toDouble(), rot.y().toDouble(), rot.z().toDouble(), rot.w().toDouble())
 
             val keyOffset = hand * 2
             updateVector2fState(thumbstickAction, handPath)
